pi-worker/requirements.txt
python-escpos>=3.0
requests>=2.28.0

